test(cli): split doctor gateway token suite
